+++ src/main/java/com/beanbeanjuice/command/settings/poll/SetPollChannelSubCommand.java
@@ -1,4 +1,4 @@
-package com.beanbeanjuice.command.moderation.poll;
+package com.beanbeanjuice.command.settings.poll;
 
 import com.beanbeanjuice.Bot;
 import com.beanbeanjuice.utility.command.CommandCategory;
@@ -10,42 +10,43 @@ import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEve
 import org.jetbrains.annotations.NotNull;
 
 /**
- * An {@link ISubCommand} used to remove the {@link Poll Poll} {@link net.dv8tion.jda.api.entities.TextChannel TextChannel}.
+ * An {@link ISubCommand} used to set the {@link Poll Poll} {@link net.dv8tion.jda.api.entities.TextChannel TextChannel}.
  *
  * @author beanbeanjuice
  */
-public class RemovePollChannelSubCommand implements ISubCommand {
+public class SetPollChannelSubCommand implements ISubCommand {
 
     @Override
     public void handle(@NotNull SlashCommandInteractionEvent event) {
 
-        // Makes sure the poll channel has been removed from the database.
-        if (GuildHandler.getCustomGuild(event.getGuild()).setPollChannel("0")) {
+        // Makes sure the poll channel has been added to the database.
+        if (GuildHandler.getCustomGuild(event.getGuild()).setPollChannel(event.getChannel().getId())) {
             event.getHook().sendMessageEmbeds(Helper.successEmbed(
-                    "Removed Poll Channel",
-                    "The poll channel has been successfully removed."
+                    "Set Poll Channel",
+                    "The current channel has been set to a `poll` channel!"
             )).queue();
             return;
         }
+
         event.getHook().sendMessageEmbeds(Helper.sqlServerError()).queue();
     }
 
     @NotNull
     @Override
     public String getDescription() {
-        return "Remove the poll channel!";
+        return "Set the current channel to poll channel!";
     }
 
     @NotNull
     @Override
     public String exampleUsage() {
-        return "`/poll-channel remove`";
+        return "`/poll-channel set #general` or `/poll-channel set`";
     }
 
     @NotNull
     @Override
     public CommandCategory getCategoryType() {
-        return CommandCategory.MODERATION;
+        return CommandCategory.SETTINGS;
     }
 
     @NotNull
@@ -57,7 +58,7 @@ public class RemovePollChannelSubCommand implements ISubCommand {
     @NotNull
     @Override
     public String getName() {
-        return "remove";
+        return "set";
     }
 
 }
