+++ src/main/java/com/beanbeanjuice/command/settings/welcome/WelcomeChannelCommand.java
@@ -1,4 +1,4 @@
-package com.beanbeanjuice.command.moderation.raffle;
+package com.beanbeanjuice.command.settings.welcome;
 
 import com.beanbeanjuice.utility.command.CommandCategory;
 import com.beanbeanjuice.utility.command.ICommand;
@@ -11,11 +11,10 @@ import org.jetbrains.annotations.Nullable;
 import java.util.ArrayList;
 
 /**
- * An {@link ICommand} used to set/remove the {@link com.beanbeanjuice.utility.section.moderation.raffle.Raffle Raffle} {@link net.dv8tion.jda.api.entities.TextChannel TextChannel}.
- *
- * @author beanbeanjuice
+ * An {@link ICommand} used for setting/removing the welcome {@link net.dv8tion.jda.api.entities.TextChannel TextChannel}
+ * in a {@link net.dv8tion.jda.api.entities.Guild Guild} and/or editing the welcome message.
  */
-public class RaffleChannelCommand implements ICommand {
+public class WelcomeChannelCommand implements ICommand {
 
     @Override
     public void handle(@NotNull SlashCommandInteractionEvent event) { }
@@ -23,27 +22,28 @@ public class RaffleChannelCommand implements ICommand {
     @NotNull
     @Override
     public String getDescription() {
-        return "Set or remove the raffle channel!";
+        return "Set/Remove the welcome channel or edit the welcome message!";
     }
 
     @NotNull
     @Override
     public String exampleUsage() {
-        return "`/raffle-channel set` or `/raffle-channel remove`";
+        return "`/welcome-channel set` or `/welcome-channel remove` or `/welcome-channel edit-message`";
     }
 
     @NotNull
     @Override
     public CommandCategory getCategoryType() {
-        return CommandCategory.MODERATION;
+        return CommandCategory.SETTINGS;
     }
 
     @NotNull
     @Override
     public ArrayList<ISubCommand> getSubCommands() {
         ArrayList<ISubCommand> subCommands = new ArrayList<>();
-        subCommands.add(new SetRaffleChannelSubCommand());
-        subCommands.add(new RemoveRaffleChannelSubCommand());
+        subCommands.add(new SetWelcomeChannelSubCommand());
+        subCommands.add(new RemoveWelcomeChannelSubCommand());
+        subCommands.add(new EditWelcomeMessageSubCommand());
         return subCommands;
     }
