+++ src/main/java/com/beanbeanjuice/utility/command/CommandCategory.java
@@ -11,7 +11,8 @@ public enum CommandCategory {
     SOCIAL("Commands used for social media stuff", "https://cdn.beanbeanjuice.com/images/cafeBot/category_type/social.gif"),
     INTERACTION("Commands used for user interactions.", "https://cdn.beanbeanjuice.com/images/cafeBot/category_type/interaction.png"),
     TWITCH("Commands used for twitch.", "https://cdn.beanbeanjuice.com/images/cafeBot/category_type/twitch.jpg"),
-    MODERATION("Commands used for moderation.", "https://cdn.beanbeanjuice.com/images/cafeBot/category_type/moderation.png");
+    MODERATION("Commands used for moderation.", "https://cdn.beanbeanjuice.com/images/cafeBot/category_type/moderation.png"),
+    SETTINGS("Commands used for bot settings.", "https://cdn.beanbeanjuice.com/images/cafeBot/category_type/settings.gif");
 
     private final String message;
     private final String link;
