+++ src/main/java/com/beanbeanjuice/utility/command/CommandHandler.java
@@ -6,16 +6,16 @@ import com.beanbeanjuice.command.fun.*;
 import com.beanbeanjuice.command.fun.birthday.BirthdayCommand;
 import com.beanbeanjuice.command.games.*;
 import com.beanbeanjuice.command.generic.*;
-import com.beanbeanjuice.command.moderation.poll.AddPollCommand;
+import com.beanbeanjuice.command.moderation.AddPollCommand;
 import com.beanbeanjuice.command.interaction.*;
 import com.beanbeanjuice.command.moderation.ClearChatCommand;
 import com.beanbeanjuice.command.moderation.CreateEmbedCommand;
-import com.beanbeanjuice.command.moderation.ListCustomChannelsCommand;
-import com.beanbeanjuice.command.moderation.counting.CountingChannelCommand;
-import com.beanbeanjuice.command.moderation.poll.PollChannelCommand;
-import com.beanbeanjuice.command.moderation.raffle.AddRaffleCommand;
-import com.beanbeanjuice.command.moderation.raffle.RaffleChannelCommand;
-import com.beanbeanjuice.command.moderation.welcome.WelcomeChannelCommand;
+import com.beanbeanjuice.command.settings.ListCustomChannelsCommand;
+import com.beanbeanjuice.command.settings.counting.CountingChannelCommand;
+import com.beanbeanjuice.command.settings.poll.PollChannelCommand;
+import com.beanbeanjuice.command.moderation.AddRaffleCommand;
+import com.beanbeanjuice.command.settings.raffle.RaffleChannelCommand;
+import com.beanbeanjuice.command.settings.welcome.WelcomeChannelCommand;
 import com.beanbeanjuice.command.social.CountMembersCommand;
 import com.beanbeanjuice.utility.logging.LogLevel;
 import net.dv8tion.jda.api.JDA;
